Teach reader to remember the last match offset
If a search string does not match anymore (because of an edit), this
offset stays around, so we can still know where the search string
is located.

This will also be handy to avoid needless cursor jumping.
Example: the cursor is at the end of the search string (marked by ^s)

echo 12 123
        ^^^

Pressing backspace results in the command line

echo 12 123
        ^^

Because we remember the match position and try that first.
Without this we'd jump back to the first occurrence of 12.