Teach {up,down}-or-search to only search when the cursor is at the search string
Since we do not stop history search on commands like cursor movements,
before this patch it was impossible to use the up and down keys to move
between lines in a multi-line command buffer.  Do the line movement if
the cursor is not at the matched search string, otherwise search.