Enable interactive mode when running batch commands
For a long time we have fought a subtle case of fish somehow not
reporting/setting the expected working directory when opening a Vim
session interactively vs. from command mode. It can be reproduced as
follows:

- Working:
  fish
  nvim
  :cd $foodir
  :!tmux display-message -p "\#{pane_current_path}"
  > $foodir

- Not working:
  fish -c nvim
  :cd $foodir
  :!tmux display-message -p "\#{pane_current_path}"
  > $HOME

While the issue is openly not fully understood, the main difference
between the working and the not working case is that one sets the
interactivity flag. This in turn has implications for job control and
process group handling, but it is not 100% clear which of these has what
effect and why.
With this change we try better to determine interactivity setting the
interactivity flag on the path used when the -c option is provided to
fix the issue.

Edit:
Okay, this issue is now better understood. `tmux display-message -p
"\#{pane_current_path}"` looks the process group's working directory.
Process groups are created differently for foreground vs. background
processes or interactive vs. non-interactive. *That's* the difference.
Take a look at:
  $ pstree -g 16152 -p

So what I ended up doing was to stop having fish as the login shell and
then using `[ -z "${NOFISH:-}" ] && SHELL=/bin/fish exec fish` in my
~/.bashrc. That's the recommended setup that Gentoo calls out as well.
That's happening after the interactivity check in the same file. The
impact that has, I suspect, is that now we actually run as bash when I
use Win-V to spawn an editor instance directly. Though..I don't think
that's how it works either..