Enable interactive mode when running batch commands
For a long time we have fought a subtle case of fish somehow not
reporting/setting the expected working directory when opening a Vim
session interactively vs. from command mode. It can be reproduced as
follows:

- Working:
  fish
  nvim
  :cd $foodir
  :!tmux display-message -p "\#{pane_current_path}"
  > $foodir

- Not working:
  fish -c nvim
  :cd $foodir
  :!tmux display-message -p "\#{pane_current_path}"
  > $HOME

While the issue is openly not fully understood, the main difference
between the working and the not working case is that one sets the
interactivity flag. This in turn has implications for job control and
process group handling, but it is not 100% clear which of these has what
effect and why.
With this change we try better to determine interactivity setting the
interactivity flag on the path used when the -c option is provided to
fix the issue.